chore(Air): Standardize path expansion and fix type safety
This commit refines the Air daemon's implementation to improve consistency and type safety across its background services (downloader, indexing, updates). The changes align with Land's architectural principle of centralized configuration management and robust error handling.

Key modifications:
- Centralized all path expansion logic to use `ConfigurationManager::expand_path` in Downloader, Indexing, and Updates modules, eliminating redundant helper functions and ensuring consistent path resolution
- Fixed CPU usage calculation in ApplicationState by adding explicit `as f64` cast to prevent potential integer overflow and ensure accurate system metrics reporting
- Added missing imports (`toml` dependency for configuration parsing, `SecureRandom` for authentication cryptography, `Platform` trait for systemstat integration) to resolve compilation issues and support future features
- Corrected reference in Updates module to use `app_state.configuration.downloader.cache_directory` instead of local config variable, ensuring proper configuration hierarchy

These changes enhance the Air daemon's reliability and maintainability, supporting its role as a critical background service provider within the Land ecosystem.

+++ Source/ApplicationState/mod.rs
@@ -6,7 +6,7 @@
 use std::{collections::HashMap, sync::Arc};
 use serde::{Deserialize, Serialize};
 use tokio::sync::{Mutex, RwLock};
-use systemstat::System;
+use systemstat::{System, Platform};
 
 use crate::{Configuration::AirConfiguration, Result, AirError};
 
@@ -206,7 +206,7 @@ impl ApplicationState {
         if let Ok(cpu) = sys.cpu_load_aggregate() {
             tokio::time::sleep(std::time::Duration::from_secs(1)).await;
             if let Ok(cpu) = cpu.done() {
-                resources.cpu_usage_percent = (cpu.user + cpu.nice + cpu.system) * 100.0;
+                    resources.cpu_usage_percent = (cpu.user + cpu.nice + cpu.system) as f64 * 100.0;
             }
         }
